requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
attrs==26.1.0
bitarray==3.8.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.7
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cssselect2==0.9.0
cycler==0.12.1
cytoolz==1.1.0
Django==5.2.11
dotenv==0.9.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Flask==2.3.0
fonttools==4.61.1
freetype-py==2.5.1
frozenlist==1.8.0
gunicorn==25.1.0
hexbytes==1.3.1
hmmlearn==0.3.3
html5lib==1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.1
numpy==2.4.2
oscrypto==1.3.0
packaging==26.0
pandas==3.0.0
parsimonious==0.10.0
patsy==1.0.2
pillow==12.1.1
propcache==0.4.1
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdf==6.7.1
python-bidi>=0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
rlp==4.1.0
rlPyCairo==0.4.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.5
statsmodels==0.14.6
svglib==1.6.0
threadpoolctl==3.6.0
tinycss2==1.5.1
toolz==1.1.0
types-requests==2.32.4.20260324
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
web3==7.14.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.3.0
whitenoise==6.11.0
xhtml2pdf==0.2.17
yarl==1.23.0
djangorestframework==3.15.1
django-cors-headers==4.3.1
